+++ comb_spec_searcher/comb_spec_searcher.py
@@ -145,7 +145,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             for start_label, end_labels, rule in self._expand_class_with_strategy(
                 comb_class, strategy, label
             ):
-                self._add_rule(start_label, end_labels, rule)
+                self.add_rule(start_label, end_labels, rule)
 
     def _expand(
         self,
@@ -165,7 +165,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                 for start_label, end_labels, rule in self._expand_class_with_strategy(
                     comb_class, strategy_generator, label
                 ):
-                    self._add_rule(start_label, end_labels, rule)
+                    self.add_rule(start_label, end_labels, rule)
 
     @staticmethod
     def _rules_from_strategy(
@@ -258,15 +258,15 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                     )
             yield start_label, tuple(end_labels), rule
 
-    def _add_rule(
+    @cssmethodtimer("add rule")
+    def add_rule(
         self, start_label: int, end_labels: Tuple[int, ...], rule: AbstractRule
     ) -> None:
         """
-        Add the cleaned rules labels.
+        Add the rule to the searcher
 
         - try to verify children combinatorial classes
         - set workability of combinatorial classes
-        - remove empty combinatorial classes
         - symmetry expand combinatorial classes
         - add class to classqueue
         """
@@ -323,7 +323,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             ):
                 inf_class = rule.children[0]
                 inf_label = end_labels[0]
-                self._add_rule(start_label, end_labels, rule)
+                self.add_rule(start_label, end_labels, rule)
                 self.classqueue.set_not_inferrable(start_label)
                 inferral_strategies = (
                     inferral_strategies[i + 1 :] + inferral_strategies[0 : i + 1]
@@ -529,6 +529,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             spec_search_start = time.time()
             logger.debug("Searching for specification.")
             if self.has_specification():
+                logger.info("Specification detected.")
                 return self.ruledb.get_specification_rules(
                     smallest=smallest,
                     minimization_time_limit=0.01 * (time.time() - auto_search_start),