+++ comb_spec_searcher/rule_db/__init__.py
@@ -1,4 +1,5 @@
 from .base import RuleDB
+from .forest import RuleDBForest
 from .forget import RuleDBForgetStrategy
 
-__all__ = ["RuleDB", "RuleDBForgetStrategy"]
+__all__ = ["RuleDB", "RuleDBForgetStrategy", "RuleDBForest"]