+++ comb_spec_searcher/strategies/constructor/cartesian.py
@@ -539,10 +539,10 @@ class Quotient(Constructor[CombinatorialClassType, CombinatorialObjectType]):
 
     @staticmethod
     def _poly_to_terms(poly: Union[sympy.Poly, int]) -> Terms:
-        if isinstance(poly, int):
+        if isinstance(poly, (int, sympy.core.numbers.Integer)):
             if poly == 0:
                 return Counter()
-            return Counter({tuple(): poly})
+            return Counter({tuple(): int(poly)})
         return Counter(poly.as_dict())
 
     def _b(
@@ -582,16 +582,15 @@ class Quotient(Constructor[CombinatorialClassType, CombinatorialObjectType]):
     def get_equation(
         self, lhs_func: sympy.Function, rhs_funcs: Tuple[sympy.Function, ...]
     ) -> sympy.Eq:
-        res = lhs_func.subs(self.extra_parameters[self.idx])
-        for (idx, rhs_func), extra_parameters in zip(
-            enumerate(rhs_funcs), self.extra_parameters
-        ):
-            if self.idx != idx:
-                res /= rhs_func.subs(
-                    {child: parent for parent, child in extra_parameters.items()},
-                    simultaneous=True,
-                ).subs(self.extra_parameters[self.idx], simultaneous=True)
-        return sympy.Eq(rhs_funcs[self.idx], res)
+        if any(self.extra_parameters):
+            raise NotImplementedError(
+                "Quotient equation is not implemented with extra parameters. "
+                "You can fall back on the cartesian equation."
+            )
+        res = rhs_funcs[0]
+        for rhs_func in rhs_funcs[1:]:
+            res /= rhs_func
+        return sympy.Eq(lhs_func, res)
 
     def reliance_profile(self, n: int, **parameters: int) -> RelianceProfile:
         raise NotImplementedError